backend/requirements-security.txt
# Production Security Dependencies
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Security
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Rate Limiting
slowapi==0.1.9

# Input Validation
bleach==6.1.0
validators==0.22.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Database Security
sqlalchemy-utils==0.41.1

# Environment Management
python-decouple==3.8